moving query optimiszer logic in different class
